// src/enums/success-code.enum.ts
enum SuccessCode {
  Ok = 'OK',
  UserCreated = 'USER_CREATED',
  UserUpdated = 'USER_UPDATED',
  UserDeleted = 'USER_DELETED',
  CodeSent = 'CODE_SENT',
  UserLoggedIn = 'USER_LOGGED_IN',
  PasswordUpdated = 'PASSWORD_UPDATED',
  CategoryCreated = 'CATEGORY_CREATED',
  CategoryUpdated = 'CATEGORY_UPDATED',
  CategoryDeleted = 'CATEGORY_DELETED',
  ProductCreated = 'PRODUCT_CREATED',
  ProductUpdated = 'PRODUCT_UPDATED',
  ProductDeleted = 'PRODUCT_DELETED',
  AddedToCart = 'ADDED_TO_CART',
  CartToBeDeleted = 'CART_TO_BE_DELETED',
  GalleryCreated = 'GALLERY_CREATED',
  GalleryUpdated = 'GALLERY_UPDATED',
  GalleryDeleted = 'GALLERY_DELETED',
  BannerCreated = 'BANNER_CREATED',
  BannerUpdated = 'BANNER_UPDATED',
  BannerDeleted = 'BANNER_DELETED',
  LeadCreated = 'LEAD_CREATED',
  LeadUpdated = 'LEAD_UPDATED',
  LeadDeleted = 'LEAD_DELETED',
  PromoCodeCreated = 'PROMO_CODE_CREATED',
  PromoCodeUpdated = 'PROMO_CODE_UPDATED',
  OrderCreated = 'ORDER_CREATED',
  OrderStatusUpdated = 'ORDER_STATUS_UPDATED',
  AddedToWishlist = 'ADDED_TO_WISHLIST',
  RemovedFromWishlist = 'REMOVED_FROM_WISHLIST',
}

export default SuccessCode;